myproject/requirements.txt
# Django and Core Dependencies
Django>=6.0.0
python-dotenv>=1.0.0

# AI/ML Dependencies
groq>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0

# Optional: For TensorFlow models
# tensorflow>=2.13.0

# Optional: For PyTorch models
# torch>=2.0.0
# torchvision>=0.15.0

# Celery for async tasks (if needed)
# celery>=5.3.0
# redis>=4.5.0

# Database
# psycopg2-binary>=2.9.0  # For PostgreSQL
# mysqlclient>=2.2.0      # For MySQL

# Development tools
# django-debug-toolbar>=4.1.0
# pytest>=7.4.0
# pytest-django>=4.5.0
